clients/python/requirements.txt
asn1crypto==0.22.0
cffi==1.10.0
cryptography==2.0.3
idna==2.6
pycparser==2.18
pyzmq==16.0.2
six==1.10.0
